requirements.txt
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.1
xgboost
scikit-learn
numpy
pandas
joblib
tensorflow